+++ tiger-java-agent/src/main/java/de/gematik/test/tiger/TigerMasterSecretListeners.java
@@ -1,0 +1,6 @@
+package de.gematik.test.tiger;
+
+public interface TigerMasterSecretListeners {
+
+  void onMasterSecret(Object session);
+}